chore: deprecate 'jade' package instrumentation
The 'jade' package (https://www.npmjs.com/package/jade) was deprecated
and renamed to 'pug' in 2015. This change deprecates our instrumentation
of it, and stops testing it. The main benefit is that its removal from
"devDependencies" removes "1 low, 4 critical" entries from an `npm
audit`. While these issues don't affect our users, because it is a
*dev* dep, those audit warnings are noise in the way of auditing our
deps.