fix: avoid spurious warning about serviceName in Lambda (#2727)
Recently in #2684 a change was made so that the agent uses
`require.main` to look for "package.json" to infer serviceName.
That doesn't work for a Lambda Node.js env, because `require.main`
is in the Lambda Runtime ("/var/runtime/...") and not the same tree
where the lambda handler is ("/var/task/..."). The result is a spurious
log.warn in lambda logs:
    could not infer serviceName: could not find package.json up from /var/runtime

Per spec, we prefer the AWS_LAMBDA_FUNCTION_NAME envvar for serviceName
so won't use the package.json value at all. This changes to avoid
looking for package.json at all.