chore: bump to got@11.8.5 for vuln fix
got@11.8.5 includes a fix for CVE-2022-33987. Ours is a dev-dep so the
vuln shouldn't affect us. This update is effectively the same thing
as the recent #2784 from dependabot.

This also removes the change from #2785 to lock us to got@10, and
clarifies that we want to continue testing with recent versions of 'got'
as a sanity check that the agent works with it. At least for now.