drop the "exports", it can be problematic
Adding this with*out* extension would be problematic. Currently without
an "exports" a "foo.mjs" can `import 'elastic-apm-node/start.js'` (requires the .js)
and can `import 'elastic-apm-node/lib/config.js'` et al. That's fine. No
need to lock it down. If we *really* want the extension*less* access names,
then we need to add "exports"... which locks it down, which *could* be
breaking. So... let's not.
I think we should bias to including the extension as ESM becomes more popular.