fix: an S3 ListBuckets span was invalid for APM server intake (#2866)
The APM agent spec for S3 spans says
`span.context.destination.service.resource` should be the bucket name
or access point, if available. That isn't available for ListBuckets,
which doesn't operate on a bucket. In this case the APM agent was emiting:
    span.context.destination.service = { name: ..., type: ... }
which fails APM server schema validation with
    missingProperty: 'resource'
and results in APM server intake logs with:
    Error: validation error: span: context: destination: service: 'resource' required

The fix, for now, is to skip `span.context.destination.service` entirely
for ListBuckets.