use node's own emitWarning to warn about users using this deprecated method
Ideally node's warning system would offer (and even default) to showing
a given DeprecationWarning with a given code just *once* to not clutter
so much. Also... supporting user easily filtering out these warnings in
code. They are limited to using the 'warning' event *and* need to then
use `node --no-warnings`. We can always revert this if too noisy,
and make it a log.warn, say.