fix avoids a possible 'RangeError: Maximum call stack size exceeded' in maybeTime(...)
This avoids a possible 'RangeError: Maximum call stack size exceeded' as
described in #2929.

The case the user hit is in `maybeTime(...)` used in Span timer handling
by dropping the recursion.  A `Timer` instance's `_timer` is always the
same object as its parent's `_timer` if it has a parent, so there is no
need to walk the chain.

Refs: #2929