test: drop the 'transBeforeClient' testing for redis >=4 tests
The 'transBeforeClient' part of tests for redis <4 were to ensure the
instrumentation was not creating *two* spans as a side-effect of
node-redis's queuing commands before "ready" and calling
`RedisClient.internal_send_command` twice. The redis@4 instrumentation
doesn't shim any internal redis commands that are re-called for
queued commands.

This also drops a test block that was never called in redis.test.js
(for redis <=2.4.2).