test: fix tests to work with restify@9
restify@9 was recently released. It only works with node >=v14.18.0
because it added use of https://nodejs.org/api/modules.html#core-modules
This updates restify testing to skip out when there is an incompat
(restify, node) version pair.

Note: It looks like the intent may have been for restify@9 to support
node >=10.0.0, so this may be refined later.