fix TAV test failure with @elastic/elasticsearch@8.6.0 and node v14 (#3100)
The test for `error.exception.module` for a reported APM error using
`@elastic/elasticsearch` broke with the recent `8.6.0` version...  as
installed by the npm in node v14.

The issue was that the install tree changed such that the error stack
top frame was in:
    node_modules/@elastic/elasticsearch/node_modules/@elastic/transport/lib/Transport.js
rather than:
    node_modules/@elastic/transport/lib/Transport.js
for reasons independent of the `@elastic/elasticsearch@8.6.0` release.
The `_moduleNameFromFrames` utility didn't handle a nested
`node_modules/...` dir.  This change fixes that.