test: fix TAV=@hapi/hapi test failure with node v8
The 'http.request' instrumentation fix in #3090 uncovered a bug in our
tests for @hapi/hapi when with node v8 (a version before
`http.request(url, opts, cb)` was supported in node). This fixes that.

This also updates the .tav.yml section for '@hapi/hapi' to align with
logic in "_is_hapi_incompat.js" so the TAV tests don't waste time
installing a version of Hapi for which tests will be skipped anyway.

Refs: #3090