rejigger to no longer use otel.metrics.setGlobalMeterProvider()
because that means a user cannot use it. Instead we instrument
otel.metrics.getMeterProvider() and provide one from the APM agent
if the user hasn't registered one. That means that user code can
use metrics from the OTel API:
    const otel = require('@opentelemetry/api')
    const meter = otel.metrics.getMeter('my-meter')
    const aCounter = meter.createCounter('a_counter')
    ...
and if the APM agent is in use, then metrics will go to APM server.