ci: do NOT fail-fast on test matrix; test windows with node 16
With `fail-fast: false` we get the useful information for whether a
particular test fails only for certain node versions (or other matrix
params) and passes on others. It can also help show if a particular
failure may be flakey (one failure, but many successes for other
versions).  The "fail-fast" default for our tests results in less useful
information

Bump Windows testing to node 16 to get closer to the current node LTS.