test: fix s3.test.js failure in Docker on macos due to timeout
It was found that running this test in Docker on mac -- e.g. by
running `.ci/scripts/test.sh -b release 18` -- would fail. It would
hit the set timeout for the `execFile` call. THis bumps the timeout
and improves some logging output around that so it is easier to
diagnose if it happens again.