tests: stop testing 'express-graphql' instrumentation
https://www.npmjs.com/package/express-graphql is deprecated.
https://github.com/graphql/express-graphql is archived.
The latest release, express-graphql@0.12.0 was 2022-11-19.

Support for instrumenting this will remain, but we will stop testing it.
A motivation for dropping this is that the `preinstall: npm uninstall express-graphql`
parts of the `graphql` TAV testing could be significantly contributing to the
graphql TAV tests being among the slowest; similar for `apollo-server-express`
TAV tests.