fix: correct supported versions of knex to exclude 0.9.0
Our knex instrumentation only ever supported '>0.9.0', but the docs for
it accidentally stated '>=0.9.0'. Then in a recent change to .tav.yml we
actually started testing with '0.9.0', the result being that the tests
hang.  This change corrects and restates the range to be '>=0.10.0
<1.0.0'.

Dropping a documented supported version might in general be worthy of
changelog, docs, and even a breaking change. However, instrumenting
0.9.0 never worked and it was released 2015.