chore: add 'npm run lint:tav'
This checks that .ci/tav.json is correct: both that it covers all the
modules in all the .tav.yml files in the repo, and that it never
results in a TAV test matrix greater than the GH Actions 256 limit.
This is part of `npm run lint` which is run by the lint CI workflow.