avoid starting the Agent in Worker threads when using start.js
This avoids the, arguably, surprise case of an additional APM agent
being started whenever a new Worker is created (see Node.js
`worker_thread` docs) if `node -r elastic-apm-node/start.js ...`
or `import 'elastic-apm-node/start.js'` is being used.

The former is especially surprising when using a loader with node v20:
    node --experimental-loader=elastic-apm-node/loader.mjs -r elastic-apm-node/start.js ...
because Node v20 has moved loader processing to a Worker... so the
"start.js" is executed again in that worker.

This change isn't ideal because an explicit use of
    import 'elastic-apm-node/start.js'
in TypeScript or ESM code for a worker does *not* start the APM agent.
That's a subtlety we will need to document.