fix: do not start the Agent in "start.js" when in a Worker thread (#3410)
This avoids the, arguably, surprise case of an additional APM agent
being started whenever a new Worker is created (see Node.js
`worker_thread` docs) if `node -r elastic-apm-node/start.js ...`
is being used.

This is especially surprising when using a loader with node v20:
    node --experimental-loader=elastic-apm-node/loader.mjs -r elastic-apm-node/start.js ...
because Node v20 has moved loader processing to a Worker... so the
"start.js" is executed again in that worker. The visible side-effect
is that the logging preamble is printed twice.

This change isn't ideal because an explicit use of
    import 'elastic-apm-node/start.js'
in TypeScript or ESM code for a worker does *not* start the APM agent.
That's a subtlety that we rely on documentation to explain away.