test: fix a flaky test that was relying on www.google.com:81 behaviour
The test was using www.google.com:81 behaviour of firewalling requests
to that port such that establishing a *connection* would hang long
enough for the code under test to emit a "connectionTimeout" event.
Recently a number of tests on unrelated PRs were getting an ECONNREFUSED
instead.

This changes the test to not rely on an external server, but uses a
custom http.Agent that drops socket "connect" events, there by
simulating a connection hang and timeout.