fix: correct dev-deps required for the recent estlin@8 upgrade
The upgrade to eslint@8 in #3409 subtly broken peer-dependencies.
`npm ls` unhelpfully does not complain, but `npm ls -a` shows some
errors, and `./dev-utils/make-distribution.sh` was broken.

The upgrade bumped "eslint-config-standard" from ^14.1.1 to ^16.
However, v16 has:
    "peerDependencies": {
      "eslint": "^7.12.1",
      "eslint-plugin-import": "^2.22.1",
      "eslint-plugin-node": "^11.1.0",
      "eslint-plugin-promise": "^4.2.1 || ^5.0.0"
    },
so is incompatible with eslint@8. We need eslint-config-standard@17,
which has:
    "peerDependencies": {
      "eslint": "^8.0.1",
      "eslint-plugin-import": "^2.25.2",
      "eslint-plugin-n": "^15.0.0 || ^16.0.0 ",
      "eslint-plugin-promise": "^6.0.0"
    },

Note that also requires a change from eslint-plugin-node (unmaintained)
to the replacement "eslint-plugin-n".