test: workaround a flaky test failure on slow Windows CI
The "end to end test with compression" would somewhat frequently fail
in CI on test-windows, e.g.:
https://github.com/elastic/apm-agent-nodejs/issues/3313#issuecomment-1634960896

This is a workaround that gives more time (200ms instead of 50ms) for the CI
machine to complete the three fast spans.  Removing the
`spanStackTraceMinDuration` copypasta should help make the spans faster as well.