chore: update deps
This updates a number of dependencies now that our base Node.js version
is v14.

There is some subtlety in the 'pino' and 'glob' deps. They both use
AbortController (pino via its usage of readable-stream@4) or
<AbortSignal>.reason to support aborting async operations. Full support
for these were not added until Node.js 14.17 -- yet this APM agent
currently supports back to 14.5.0.

I feel this is not an issue. In node 14.x `AbortController` et al are only available
behind a flag (`node --experimental-abortcontroller`).  Node.js v14.17
added `<AbortSignal>.reason`, so the fail case with Pino is a user using
(a) a Node.js version in the range [v14.5, v14.17), and (b) with the
`--experimental-abortcontroller` flag, and (c) something in the APM
agent's code path tries to use `AbortSignal.reason`. When the flag is
not used both readable-stream@4 and glob@10 will use a polyfill.

We could consider setting min to >14.17, but I don't currently think it
is necessary.

Closes: https://github.com/elastic/apm-agent-nodejs/issues/2760