LOCALSTACK_HOST=localstack:4566 and SQS_ENDPOINT_STRATEGY=path
IIUC, since localstack/localstack:2.3.0 the default LOCALSTACK_HOST
is `localhost.localstack.cloud:4566` which relies on *external* DNS
to resolve `*.localhost.localstack.cloud` hostnames -- as real AWS
server and client tools do for some services like S3 and SQS.

We can't rely on this external DNS access for our in-docker tests,
i.e. for the current TAV tests. We avoid using this with our
LOCALSTACK_HOST setting and tell the (apparently server-side) handling
for SQS queue URLs to not use hostname-based queue URLs.

This is similar to path vs host style for S3, which is apparently
client-side determined -- again IIUC from
https://docs.localstack.cloud/user-guide/aws/s3/#path-style-and-virtual-hosted-style-requests