test: standardize on shorter healthcheck interval in docker-compose test config (#4169)
For some testing (GitHub Actions test.yml is separate), we use test/docker-compose.yml
for test services. This standardizes on this healthcheck timing:

      interval: 1s
      timeout: 10s
      retries: 30

The previous 'interval: 30s' for some services was unnecessarily slow.
It means the first healthcheck would always wait that long and one would
be waiting doing nothing. It is possibel with this change we will have a
very slow service not come up in time. If so, then we can increase the
retries or interval. Preceding values were willy-nilly.