correct the 'Error.prepareStackTrace is not set' test assumptions, node v21.6.0 sets a default Error.prepareStackTrace (#4295)
This cherry-picks commit b8daf0385 from the main branch (actually one of the squashed commits in PR 3827):
https://github.com/elastic/apm-agent-nodejs/pull/3827/commits/b8daf0385e5b40dd7df28ce9d8230225608bcac1

This fixes an error in stacktraces.test.js when running with Node.js 20. That change to node v21.6.0 also made it into v20.12.0, which means this test started breaking when that was released (2024-03-26).