test: test FIPS 140 compliance (#4441)
Adds a 'test-fips' CI workflow that is run weekly. Complains to Slack if it fails.

Also:
This simplifies down from 3 different usages of a generated cert in
a number of test files. We also update the generated cert to be RSA
2048 for FIPS 140.  We can also now drop the https-pem devDep
which was generating a cert in a post-install script, which is
sometimes painful with supply-chain security efforts.