test: fix testing for @elastic/elasticsearch version 9 (#4583)
`@elastic/elasticsearch@9.0.0` was released recently. This updates our testing to pass with it. So far no instrumentation changes were necessary. (I'm not sure if testing with v9 of the client would have failed with the older 8.7.1 version of the server that we had been using.)

Also reduce the number of ES client versions tested.

Obsoletes: #4576