chore: clarify the latest major version of the instrumentation module for some instrumentations
In docs, instrumentation, and tests, the major version of the target
module being instrumented was not specified. There should be no functional
change for any current releases of modules with this change. In other
words, any changes to actual instrumentations is capping the major version
supported at the *currently released* major version.

Closes: https://github.com/elastic/apm-agent-nodejs/issues/4803