tests,docs: stop testing against Restify
Also remove the Restify-specific getting started doc. Restify no
longer gets maintenance and has not for a long while. It doesn't work
with that latest version of Node.js. Let's not encourage its use.
As well, because it doesn't get maintenance, having the latest restify
installed as a dev-dep locks us into carrying old deps with dependabot
security warnings in them. That is noise for dep update hygiene.

This does not change the instrumentation, so it isn't a breaking change.